modification fonctionnel sans fuite

+++ src/alteration.c
@@ -9,7 +9,7 @@ int genNumber() {
 }
 
 int testDel() {
-	if(genNumber() <10) {
+	if(genNumber() <15) {
 		return 1;
 	}
 	return 0;
@@ -90,29 +90,40 @@ void deleteSommet(Liste* liste, int ligne) {
 	
 	Sommet* tmp = liste->first;
 	Sommet* prec = liste->first;
+	
 	//on trouve le sommet à supprimer
-	while (tmp->numLigne != ligne && tmp->suivant!=NULL) {
+	while (tmp->numLigne<ligne && tmp->suivant!=NULL) {
+		//~ printf("num ligne :%d\n",tmp->numLigne);
 		prec = tmp;
 		tmp = tmp->suivant;
 	}
 	
 	if (tmp->numLigne == ligne) {
+		//si c'est le premier
+		if (tmp == liste->first) {
+			//~ printf("PREMIER PAS FINI\n");
+			liste->first = tmp->suivant;
+			free(tmp);
+		}
 		//pas arrivé à la fin
-		if (tmp->suivant!= NULL) {
+		else if (tmp->suivant!= NULL) {
+			//~ printf("PAS A LA FIN\n");
 			prec->suivant = tmp->suivant;
-			tmp->suivant = NULL;
 			free(tmp);
 		}
 		//arrivé à la fin mais pas qu'un seul élément ds la liste
 		else if (tmp->suivant == NULL 
 					&& tmp!=liste->first) {
+			//~ printf("FIN\n");
 			liste->last = prec;
+			liste->last->suivant = NULL;
 			free(tmp);
 			
 		}
 		//qu'un seul élément dans la liste
 		else if (tmp->suivant == NULL 
 					&& tmp==liste->first) {
+			//~ printf("PREMIER\n");
 			liste->last = NULL;
 			liste->first = NULL;
 			free(tmp);
@@ -131,7 +142,9 @@ int delLigne2(Liste *tab, int ligne, int taille) {
 	}
 	
 	for (int i = 0; i<taille; i++) {
+		if (tab[i].exist == 1) {
 		deleteSommet(&tab[i],ligne);
+		}
 	}
 	
 	return 0;
@@ -150,7 +163,7 @@ int delColumn(Liste* tab, int size) {
 	int j = 0;
 	for(int i = 0; i < size; i++) {
 		if(testDel()) {
-			printf("Delete column %d.\n",i);
+			//~ printf("Delete column %d.\n",i);
 			j++;
 			freeListe(&tab[i]);
 			tab[i].first = NULL;
